Conhecendo as funções de alta ordem
