+++ bytebank/src/main/kotlin/br/com/alura/bytebank/teste/testaTipoFuncao.kt
@@ -1,0 +1,50 @@
+package br.com.alura.bytebank.teste
+
+fun testaFuncaoAnonima() {
+    val minhaFuncaoAnonima = fun(a: Int, b: Int): Int {
+        return a + b
+    }
+    println(minhaFuncaoAnonima(10, 20))
+
+
+    val calculaBonificacaoAnonima: (salario: Double) -> Double = fun(salario): Double {
+        if (salario > 1000.0) {
+            return salario + 50.0
+        }
+        return salario + 100.0
+    }
+    println(calculaBonificacaoAnonima(1150.0))
+}
+
+fun testaFuncaoLambda() {
+    val minhaFuncaoLambda = { a: Int, b: Int ->
+        a + b
+    }
+    println(minhaFuncaoLambda(15, 10))
+
+    val calculaBonificacao: (salario: Double) -> Double = lambda@{ salario ->
+        if (salario > 1000.0) {
+            return@lambda salario + 50.0
+        }
+        return@lambda salario + 100.0
+    }
+    println(calculaBonificacao(1100.0))
+}
+
+fun testaTipoFuncaoClasse() {
+    val minhaFuncaoClasse: (Int, Int) -> Int = Soma()
+    println(minhaFuncaoClasse(7, 3))
+}
+
+fun testaTipoFuncaoReferencia() {
+    val minhaFuncao: (Int, Int) -> Int = ::soma
+    println(minhaFuncao(2, 3))
+}
+
+fun soma(a: Int, b: Int): Int {
+    return a + b
+}
+
+class Soma : (Int, Int) -> Int {
+    override fun invoke(a: Int, b: Int): Int = a + b
+}
